refactor: Extract versioning and AppendRemoveItem into submodules
